+++ src/di/container-config.ts
@@ -10,6 +10,7 @@ import { MenuObserver } from '../ui/MenuObserver';
 import { ContextMenuHandler } from '../ui/ContextMenuHandler';
 import { SeekHandler } from '../handlers/SeekHandler';
 import { LikeButtonHandler } from '../handlers/LikeButtonHandler';
+import { DocumentFocusHandler } from '../handlers/DocumentFocusHandler';
 import { PiPWindowHandlers } from '../core/PiPWindowHandlers';
 import { PiPManager } from '../core/PiPManager';
 import { YtActionSender } from '../core/YtActionSender';
@@ -30,6 +31,7 @@ export function createContainer(): Container {
   container.bind(SeekHandler).toSelf();
   container.bind(YtActionSender).toSelf();
   container.bind(LikeButtonHandler).toSelf();
+  container.bind(DocumentFocusHandler).toSelf();
   container.bind(PiPWindowHandlers).toSelf();
   container.bind(PiPManager).toSelf();
   container.bind(MediaSessionHandler).toSelf();