+++ src/selectors.ts
@@ -11,7 +11,7 @@ export const SELECTORS = {
   YT_DRAGGABLE: 'yt-draggable',
   CONTEXT_MENU: '.ytp-popup.ytp-contextmenu',
   CONTEXT_MENU_CONTAINER: '.ytp-contextmenu',
-  PANEL_MENU_ITEMS: '.ytp-panel-menu .ytp-menuitem',
+  PANEL_MENU_ITEMS: '.ytp-panel-menu > .ytp-menuitem',
   MENU_BUTTON: '.yt-spec-button-shape-next[aria-expanded]',
   PLAYLIST_PANEL: '.ytdMiniplayerComponentPlaylistPanel',
   PROGRESS_BAR: '.ytp-progress-bar',